Fix #139
Changed the code to check for .dat before trying to configure printers, if one cannot be found an error is logged.
Removed FileException throw & fixed Console.Writeline typo.
Added a check for & and |, not the most robust method but better then allowing completely unsanitized input. To allow for any possible future debugging, if such input is detected it is changed to the `` characters, the printer is not configured and an error is logged.